+++ packages/activity/src/internals.ts
@@ -1,4 +1,7 @@
-import { AsyncLocalStorage, Context } from './types';
+export interface AsyncLocalStorage<T> {
+  getStore(): T | undefined;
+  run<A extends any[], R>(store: T, callback: (...args: A) => Promise<R>, ...args: A): Promise<R>;
+}
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 let { AsyncLocalStorage } = require('async_hooks');
@@ -10,7 +13,7 @@ if (AsyncLocalStorage === undefined) {
 /**
  * Workaround for the conditional require making AsyncLocalStorage 'any'
  */
-function getAsyncLocalStorage(): AsyncLocalStorage<Context> {
+function getAsyncLocalStorage(): AsyncLocalStorage<any> {
   return new AsyncLocalStorage();
 }
