+++ scripts/gen-docs.js
@@ -39,7 +39,7 @@ async function genDocs(package) {
 async function main() {
   let packages = process.argv.slice(2);
   if (packages.length === 0) {
-    packages = ['worker', 'client'];
+    packages = ['worker', 'client', 'activity', 'workflow'];
   }
   for (const package of packages) {
     await genDocs(package);
